+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Document new slurm-operator Token CR and add it to hack/kind.sh
+
 ### Removed
 
 - Removed the need to specify --bridge when running hack/kind.sh